requirements.txt
streamlit==1.46.0
langchain==0.2.5
langchain-community==0.2.5
langchain-google-genai==1.0.5
langchain-huggingface==0.0.3
pypdf==4.2.0
sentence-transformers==2.6.1
google-generativeai==0.5.4
faiss-cpu
pymupdf==1.24.5
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
cachetools==5.5.0
torch==2.2.2+cpu
huggingface_hub==0.23.1
--extra-index-url https://download.pytorch.org/whl/cpu